Issue #25735 separate test applications for EE 11 vs EE 10 tests
