+++ dev/com.ibm.ws.concurrent_fat_cdi/.classpath
@@ -4,6 +4,7 @@
 	<classpathentry kind="src" path="test-applications/concurrentCDIApp/src"/>
 	<classpathentry kind="src" path="test-applications/concurrentCDI2App/src"/>
 	<classpathentry kind="src" path="test-applications/concurrentCDI3App/src"/>
+	<classpathentry kind="src" path="test-applications/concurrentCDI4App/src"/>
 	<classpathentry kind="con" path="aQute.bnd.classpath.container"/>
 	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8"/>
 	<classpathentry kind="output" path="bin"/>
