+++ dev/com.ibm.ws.concurrent_fat_cdi/bnd.bnd
@@ -1,5 +1,5 @@
 #*******************************************************************************
-# Copyright (c) 2017,2022 IBM Corporation and others.
+# Copyright (c) 2017,2023 IBM Corporation and others.
 # All rights reserved. This program and the accompanying materials
 # are made available under the terms of the Eclipse Public License 2.0
 # which accompanies this distribution, and is available at
@@ -17,7 +17,8 @@ src: \
 	fat/src,\
 	test-applications/concurrentCDIApp/src,\
 	test-applications/concurrentCDI2App/src,\
-	test-applications/concurrentCDI3App/src
+	test-applications/concurrentCDI3App/src,\
+	test-applications/concurrentCDI4App/src
 
 fat.project: true
