+++ dev/com.ibm.ws.concurrent_fat_cdi/fat/src/com/ibm/ws/concurrent/cdi/fat/ConcurrentCDI4Test.java
@@ -24,23 +24,23 @@ import componenttest.annotation.TestServlet;
 import componenttest.custom.junit.runner.FATRunner;
 import componenttest.topology.impl.LibertyServer;
 import componenttest.topology.utils.FATServletClient;
-import concurrent.cdi.web.ConcurrentCDIServlet;
+import concurrent.cdi4.web.ConcurrentCDI4Servlet;
 
 @RunWith(FATRunner.class)
 @MinimumJavaLevel(javaLevel = 11)
 public class ConcurrentCDI4Test extends FATServletClient {
 
-    public static final String APP_NAME = "concurrentCDIApp";
+    public static final String APP_NAME = "concurrentCDI4App";
 
     @Server("concurrent_fat_cdi4")
-    @TestServlet(servlet = ConcurrentCDIServlet.class, contextRoot = APP_NAME)
+    @TestServlet(servlet = ConcurrentCDI4Servlet.class, contextRoot = APP_NAME)
     public static LibertyServer server;
 
     @BeforeClass
     public static void setUp() throws Exception {
-        ShrinkHelper.defaultDropinApp(server, APP_NAME, "concurrent.cdi.web");
+        ShrinkHelper.defaultDropinApp(server, APP_NAME, "concurrent.cdi4.web");
         server.startServer();
-        runTest(server, APP_NAME + '/' + ConcurrentCDIServlet.class.getSimpleName(), "initTransactionService");
+        runTest(server, APP_NAME + '/' + ConcurrentCDI4Servlet.class.getSimpleName(), "initTransactionService");
     }
 
     @AfterClass
