+++ dev/com.ibm.ws.concurrent_fat_cdi/fat/src/com/ibm/ws/concurrent/cdi/fat/ConcurrentCDITest.java
@@ -21,26 +21,33 @@ import com.ibm.websphere.simplicity.ShrinkHelper;
 import componenttest.annotation.MinimumJavaLevel;
 import componenttest.annotation.Server;
 import componenttest.annotation.TestServlet;
+import componenttest.annotation.TestServlets;
 import componenttest.custom.junit.runner.FATRunner;
 import componenttest.topology.impl.LibertyServer;
 import componenttest.topology.utils.FATServletClient;
 import concurrent.cdi.web.ConcurrentCDIServlet;
+import concurrent.cdi4.web.ConcurrentCDI4Servlet;
 
 @RunWith(FATRunner.class)
 @MinimumJavaLevel(javaLevel = 17)
 public class ConcurrentCDITest extends FATServletClient {
 
     public static final String APP_NAME = "concurrentCDIApp";
+    public static final String APP_NAME_EE10 = "concurrentCDI4App";
 
     @Server("concurrent_fat_cdi")
-    @TestServlet(servlet = ConcurrentCDIServlet.class, contextRoot = APP_NAME)
+    @TestServlets({
+                    @TestServlet(servlet = ConcurrentCDIServlet.class, contextRoot = APP_NAME),
+                    @TestServlet(servlet = ConcurrentCDI4Servlet.class, contextRoot = APP_NAME_EE10)
+    })
     public static LibertyServer server;
 
     @BeforeClass
     public static void setUp() throws Exception {
         ShrinkHelper.defaultDropinApp(server, APP_NAME, "concurrent.cdi.web");
+        ShrinkHelper.defaultDropinApp(server, APP_NAME_EE10, "concurrent.cdi4.web");
         server.startServer();
-        runTest(server, APP_NAME + '/' + ConcurrentCDIServlet.class.getSimpleName(), "initTransactionService");
+        runTest(server, APP_NAME_EE10 + '/' + ConcurrentCDI4Servlet.class.getSimpleName(), "initTransactionService");
     }
 
     @AfterClass
