+++ dev/com.ibm.ws.concurrent_fat_cdi/publish/servers/concurrent_fat_cdi4/server.xml
@@ -32,6 +32,6 @@
     </managedScheduledExecutorService>
 
     <!-- needed to shut down pool of unmanaged threads that is used by the application -->
-    <javaPermission codebase="${server.config.dir}/dropins/concurrentCDIApp.war" className="java.lang.RuntimePermission" name="modifyThread"/>
+    <javaPermission codebase="${server.config.dir}/dropins/concurrentCDI4App.war" className="java.lang.RuntimePermission" name="modifyThread"/>
 
 </server>
