+++ dev/com.ibm.ws.concurrent_fat_cdi/test-applications/concurrentCDI4App/resources/WEB-INF/web.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<web-app version="3.0"
+    xmlns="http://java.sun.com/xml/ns/javaee"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="http://java.sun.com/xml/ns/javaee web-app_3_0.xsd"
+    id="WebApp_ID">
+
+  <display-name>Concurrent FAT</display-name>
+  <description>FAT for EE Concurrency Utilities</description>
+
+  <!-- ENV ENTRIES -->
+  <env-entry>
+    <env-entry-name>java:comp/env/entry1</env-entry-name>
+    <env-entry-type>java.lang.String</env-entry-type>
+    <env-entry-value>value1</env-entry-value>
+  </env-entry>
+
+</web-app>
