+++ dev/com.ibm.ws.concurrent_fat_cdi/test-applications/concurrentCDI4App/src/concurrent/cdi4/web/TransactionalBean.java
@@ -1,16 +1,16 @@
 /*******************************************************************************
- * Copyright (c) 2017,2021 IBM Corporation and others.
+ * Copyright (c) 2017,2023 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-2.0/
- * 
+ *
  * SPDX-License-Identifier: EPL-2.0
  *
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package concurrent.cdi.web;
+package concurrent.cdi4.web;
 
 import java.io.Serializable;
 import java.util.concurrent.CompletableFuture;
